requirements.txt
aiosqlite==0.17.0
alembic==1.7.7
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
bcrypt==4.1.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.0
fastapi==0.78.0
fastapi-users==13.0.0
fastapi-users-db-sqlalchemy==4.0.5
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.6
makefun==1.15.2
Mako==1.3.2
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
pwdlib==0.2.0
pycparser==2.21
pydantic==1.10.14
PyJWT==2.8.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
sniffio==1.3.1
SQLAlchemy==1.4.36
starlette==0.19.1
typing_extensions==4.10.0
uvicorn==0.17.6
watchgod==0.8.2
websockets==12.0
